Move upgraderoom tests into their own file and add a couple
